tests/wpt/web-platform-tests/html/semantics/popovers/popover-defaultopen-2.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel=author href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div popover id=p1 defaultopen>
  Outer popover
  <div popover id=p2 defaultopen>
    Inner popover
  </div>
</div>

<script>
  requestAnimationFrame(() => {
    requestAnimationFrame(() => {
      test(function(){
        assert_true(p1.matches(':open'),'The first (outermost) popover should be the one that opens in this case');
        assert_false(p2.matches(':open'),'The inner popover should not be open');
      }, "The defaultopen attribute should cause only the first popover to open");
    });
  });
</script>
